requirements.txt
Flask==3.0.0
flask-cors==4.0.0
flask-login==0.6.3
python-dotenv==1.0.0
requests==2.31.0
redis==5.0.1
celery==5.3.4
gunicorn==21.2.0
PyJWT==2.8.0
cryptography==41.0.7
authlib==1.3.0
httpx==0.25.2
supabase==2.3.0
gotrue==2.1.0
werkzeug==3.0.1